refactor(blocklist): address review comments
